Hook into Phoenix telemetry for logging
This is the recommended approach and means we don't need custom plugs.
