+++ lib/nerves_hub_web/device_endpoint.ex
@@ -35,7 +35,5 @@ defmodule NervesHubWeb.DeviceEndpoint do
 
   plug(Sentry.PlugContext)
 
-  plug(NervesHubWeb.Plugs.Logger)
-
   plug(NervesHubWeb.Plugs.DeviceEndpointRedirect)
 end
